Rename task dropdown to task overview
