+++ ClientApp/src/app/components/miscellaneous/sprint-dropdown/sprint-dropdown.component.html
@@ -85,9 +85,10 @@
       [ngStyle]="{'margin-bottom': data.tasks.length === 0 ? '0' : 'var(--content-padding)'}"
     ></button>
 
-    <p-accordion [multiple]="true">
-      <task-dropdown *ngFor="let task of data.tasks" [taskData]="task"></task-dropdown>
-    </p-accordion>
+    <task-overview
+      *ngFor="let task of data.tasks"
+      [taskData]="task"
+    ></task-overview>
   </ng-template>
 </p-accordionTab>
