+++ ClientApp/src/app/components/pages/labels-page/labels-page.component.html
@@ -22,9 +22,7 @@
   </ng-template>
 </p-dropdown>
 
-<p-accordion
-  [multiple]="true"
-  [ngStyle]="{'margin-top': tasksByLabel.length === 0 ? '0' : 'var(--content-padding)'}"
->
-  <task-dropdown *ngFor="let task of tasksByLabel" [taskData]="task"></task-dropdown>
-</p-accordion>
+<task-overview
+  *ngFor="let task of tasksByLabel"
+  [taskData]="task"
+></task-overview>
